requirements.txt
opencv-python
numpy
matplotlib
DateTime
sklearn
scikit-learn
pandas
seaborn
tensorflow
Werkzeug==0.15.5
Flask-WTF
WTForms
xgboost==1.2.0
jinja2==3.0.3
itsdangerous==2.0.1
